fixed checkstyle and minor leaderboard issues

+++ app/src/main/java/com/example/a2340a_team10/model/Attempt.java
@@ -5,18 +5,18 @@ import com.example.a2340a_team10.R;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
-public class Attempt implements Comparable<Attempt>{
+public class Attempt implements Comparable<Attempt> {
     private int score;
     private int avatar;
     private String name;
     private String time;
 
     public Attempt() {
-        //obtained the attributes of the current player
+        // obtained the attributes of the current player
         int choice = Player.getPlayer().getCharacterChoice();
         this.name = Player.getPlayer().getName();
         this.score = Player.getPlayer().getScore();
-        SimpleDateFormat formatter =new SimpleDateFormat("HH:mm:ss");
+        SimpleDateFormat formatter = new SimpleDateFormat("HH:mm:ss");
         Date date = new Date(System.currentTimeMillis());
         this.time = formatter.format(date);
 
@@ -31,7 +31,7 @@ public class Attempt implements Comparable<Attempt>{
         }
     }
 
-    public float getScore() {
+    public int getScore() {
         return score;
     }
 
@@ -49,6 +49,6 @@ public class Attempt implements Comparable<Attempt>{
 
     @Override
     public int compareTo(Attempt attempt) {
-        return (int) (attempt.getScore() - score);
+        return (attempt.getScore() - score);
     }
 }
